Makefile: create *.po files using standalone rule
This is how make rules are expected to work. They generate files.
We still preserve the previous behavior that always tries to update all
files when update is called but we won't update po files unless pot file
is newer for other targets.

There is also one hack with touch of *.po files. The msgmerge won't
update file if it did no change but that makes it so make has to run the
rule every time. To prvent that we always update the modification time.

The userlists translations were removed in
1128f81897711d1a8612c6876e1755dfd9e6ea21 but were left in the Makefile.
This removes them from there as well.